+++ R/get_elixhauser_codes.R
@@ -16,7 +16,7 @@
 #' 
 #' @return A `data.frame` with the following columns:
 #' * `icdv`: Integer vector indicating if the code is from ICD-9 or ICD-10
-#' * `dx`: Integer vector.  1 if the code is a diagnostic, (ICD-9-CM, ICD-10-CM, WHO, CDC Mortality), or 0 if the code is procedural (ICD-9-PCS, ICD-10-PCS)
+#' * `dx`: Integer vector.  1 if the code is a diagnostic, (ICD-9-CM, ICD-10-CM, ICD-10-AM, ICD-10-SE, WHO, CDC Mortality), or 0 if the code is procedural (ICD-9-PCS, ICD-10-PCS)
 #' * `full_code`: Character vector with the ICD code and any relevant decimal point
 #' * `code`: Character vector with the compact ICD code omitting any relevant decimal point
 #' * `condition`: Character vector of the conditions