+++ R/is_icd.R
@@ -35,7 +35,7 @@
 #' @family ICD tools
 #' @export
 is_icd <- function(x, icdv = c(9L, 10L), dx = c(1L, 0L),
-                   src = c("cdc", "cms", "ihacpa", "who"),
+                   src = c("cdc", "cms", "ihacpa", "socialstyrelsen", "who"),
                    year, headerok = FALSE,
                    ever.assignable = missing(year),
                    warn.ambiguous = TRUE,