+++ R/medicalcoder-pkg.R
@@ -6,13 +6,13 @@
 #' ICD-10 diagnosis and procedure codes. It provides fast, dependency-free tools
 #' to look up, validate, and manipulate ICD codes, while also implementing
 #' widely used comorbidity algorithms such as Charlson, Elixhauser, and the
-#' Pediatric Complex Chronic Conditions (PCCC).
-#' Designed for portability and reproducibility, the package avoids external
-#' dependencies, requiring only R >= 3.5.0, yet offers a rich set of curated ICD
-#' code libraries from the United States' Centers for Medicare and Medicaid
-#' Services (CMS), Centers for Disease Control (CDC), the World Health
-#' Organization (WHO), and ICD-10 Australian Modification (ICD-10-AM) codes
-#' from the Independent Health and Aged Care Pricing Authority (IHACPA).
+#' Pediatric Complex Chronic Conditions (PCCC). Designed for portability and
+#' reproducibility, the package avoids external dependencies, requiring only R
+#' >= 3.5.0, yet offers a rich set of curated ICD code libraries from the United
+#' States' Centers for Medicare and Medicaid Services (CMS), Centers for Disease
+#' Control (CDC), the World Health Organization (WHO), the Independent Health
+#' and Aged Care Pricing Authority (IHACPA), and from Sweden's National Board of
+#' Health and Welfare (Socialstyrelsen).
 #'
 #' The package balances performance with elegance: its internal caching,
 #' efficient joins, and compact data structures make it practical for
@@ -32,19 +32,17 @@
 #'
 #' This design choice was made for several reasons.
 #'
-#' 1. Ease of installation:
-#'    * Only requirement is R >= 3.5.0.
-#'    * No need for external files, downloads, or other packages for the ICD database.
+#' 1. Ease of installation: * Only requirement is R >= 3.5.0. * No need for
+#' external files, downloads, or other packages for the ICD database.
 #'
-#' 2. Works well with different data paradigms
-#'    * Base R `data.frames`,
-#'    * [tidyverse](https://tidyverse.org/) `tibble`s, and
-#'    * `data.table`s from the [data.table](https://rdatatable.gitlab.io/data.table/) package.
+#' 2. Works well with different data paradigms * Base R `data.frames`, *
+#' [tidyverse](https://tidyverse.org/) `tibble`s, and * `data.table`s from the
+#' [data.table](https://rdatatable.gitlab.io/data.table/) package.
 #'
 #' One of the reasons for focusing on building a self-contained package with no
 #' need for additional namespaces is to make installation and use in a
-#' restricted and secured system easier. The author of this package routinely works
-#' on machines with extremely limited access to the world-wide web. As
+#' restricted and secured system easier. The author of this package routinely
+#' works on machines with extremely limited access to the world-wide web. As
 #' such, relying on any system dependencies or other R packages can become
 #' difficult as the machine may or may not have the needed software. So long as
 #' R >= 3.5.0 is available, medicalcoder will work.
@@ -56,13 +54,11 @@
 #'
 #' @seealso [comorbidities()], [get_icd_codes()], [is_icd()],
 #'
-#' * Vignette for working with ICD codes:
-#'   * `vignette(topic = "icd", package = "medicalcoder")`
-#' * Vignettes for applying comorbidities:
-#'   * `vignette(topic = "comorbidities", package = "medicalcoder")`
-#'   * `vignette(topic = "pccc", package = "medicalcoder")`
-#'   * `vignette(topic = "charlson", package = "medicalcoder")`
-#'   * `vignette(topic = "elixhauser", package = "medicalcoder")`
+#' * Vignette for working with ICD codes: * `vignette(topic = "icd", package =
+#' "medicalcoder")` * Vignettes for applying comorbidities: * `vignette(topic =
+#' "comorbidities", package = "medicalcoder")` * `vignette(topic = "pccc",
+#' package = "medicalcoder")` * `vignette(topic = "charlson", package =
+#' "medicalcoder")` * `vignette(topic = "elixhauser", package = "medicalcoder")`
 #'
 #' @keywords internal
 "_PACKAGE"