+++ man-roxygen/params-icd-src.R
@@ -1,4 +1,5 @@
 #' @param src Character vector of code sources. One or more of `"cdc"`, `"cms"`,
-#' `"ihacpa"`, and `"who"`. Defaults to all. CDC and CMS are from the United
-#' States (ICD-9/10-CM and ICD-9/10-PCS), IHACPA from Australia (ICD-10-AM),
-#' and the World Health Organization (ICD-10).
+#' `"ihacpa"`, `"socialstyrelsen"`, and `"who"`. Defaults to all. CDC and CMS
+#' are from the United States (ICD-9/10-CM and ICD-9/10-PCS), IHACPA from
+#' Australia (ICD-10-AM), Socialstyrelsen from Sweden (ICD-10-SE), and the
+#' World Health Organization (ICD-10).