+++ tests/test-vs-pccc_1.0.7.R
@@ -216,7 +216,7 @@ stopifnot(
 #subset(get_icd_codes(with.descriptions = TRUE), grepl("^Z94", full_code) & icdv == 10)
 
 mismatch_transplant <- old_vs_mdcr[old_vs_mdcr$transplant != old_vs_mdcr$any_transplant, ]
-stopifnot(nrow(mismatch_transplant) == 19L)
+stopifnot(nrow(mismatch_transplant) == 22L)
 stopifnot(mismatch_transplant$full_code %in% c("V42.0", "Z94.1", "Z94.2", "Z94.4", "Z94.81", "Z94.82", "Z94.83", "Z94.84"))
 stopifnot(
   all(mismatch_transplant$transplant == 0L),